24429141 login

+++ screens/LoginScreen.js
@@ -16,6 +16,7 @@ import {addPlace} from '../actions/place';
 import {addUser} from '../actions/user';
 import {WebBrowser} from 'expo';
 import {SQLite} from 'expo-sqlite';
+import Validator from 'validator';
 
 
 // create a global db for database list and last known user
@@ -41,7 +42,16 @@ class LoginScreen extends React.Component {
       places: 'b',
       db: (screenProps.databaseName) ? SQLite.openDatabase(screenProps.databaseName) : null,
       uid: 0,
+      urlInvalid: false,
+      loginError: false,
+      loginErrorMessage: '',
     }
+  }
+
+
+  handleLoginError(message = '') {
+    this.setState({'loginErrorMessage': message});
+    this.setState({'loginError': true});
   }
 
 
@@ -71,13 +81,15 @@ class LoginScreen extends React.Component {
     if (this.state.password !== false) {
       var pass = this.state.password.toLowerCase().trim();
     }
-    if (this.state.url === false) {
-      // Will need error handling/validation
+    if (this.state.url === false || !Validator.isURL(this.state.url)) {
+      this.setState({'urlInvalid': true});
       return;
 
     }
 
     var url = this.state.url.toLowerCase().trim();
+    // Remove trailing slash from url
+    url = url.replace(/\/$/, "");
 
     // Set a component URL state for now, then once login is complete set the app-wide URL
     this.setState({
@@ -85,8 +97,8 @@ class LoginScreen extends React.Component {
     }, () => {
 
       fetch(this.state.url + '/services/session/token')
-        .then((response) => response.text())
-        .then((response) => {
+          .then((response) => response.text())
+          .then((response) => {
             let Token = response;
 
             let data = {
@@ -106,40 +118,51 @@ class LoginScreen extends React.Component {
             };
 
             fetch(this.state.url + '/app/user/login.json', data)
-              .then((response) => response.json())
+                .then((response) => response.json())
                 .then((responseJson) => {
+
+                  // Check for user in response. If there's no user, the response is an error message.
+                  if (typeof responseJson.user === 'undefined') {
+                    this.handleLoginError(responseJson);
+                  } else {
                     // remove http:// from url
                     const url = this.state.url.replace(/(^\w+:|^)\/\//, '');
+
 
                     // we need to update our global user
                     globalDB.transaction(
                         tx => {
-                            tx.executeSql('delete from user;',
-                            );
+                          tx.executeSql('delete from user;',
+                          );
                         }
                     );
 
                     globalDB.transaction(
                         tx => {
-                            tx.executeSql('insert into user (siteUrl, user) values (?, ?)',
-                                [url, JSON.stringify(responseJson)],
-                                (success) => {
-                                  this._handleSiteUrlUpdate(this.state.url, responseJson.user.uid, true);
-                                },
+                          tx.executeSql('insert into user (siteUrl, user) values (?, ?)',
+                              [url, JSON.stringify(responseJson)],
+                              (success) => {
+                                this._handleSiteUrlUpdate(this.state.url, responseJson.user.uid, true);
+                              },
 
-                                (success, error) => console.log(' ')
-                            );
+                              (success, error) => {
+                                console.log('error');
+                              }
+                          );
                         }
                     );
 
-                  this.props.add(responseJson.session_name + '=' + responseJson.sessid);
-                  this.props.addUserProp(responseJson);
-                  this.props.navigation.navigate('Home')
+                    this.props.add(responseJson.session_name + '=' + responseJson.sessid);
+                    this.props.addUserProp(responseJson);
+                    this._handleLoginStatusUpdate();
+                    this.props.navigation.navigate('Home')
+                  }
 
                 })
 
 
                 .catch((error) => {
+                  this.handleLoginError('Error logging in.');
                   if (error.response) {
                     // The request was made and the server responded with a status code
                     // that falls out of the range of 2xx
@@ -159,10 +182,11 @@ class LoginScreen extends React.Component {
                 });
           })
           .catch((error) => {
+            this.handleLoginError('Error logging in.');
             // console.error(error);
           });
 
-    });
+    }, this);
   }
 
   render() {
@@ -170,6 +194,17 @@ class LoginScreen extends React.Component {
     let showError = [];
     if (this.state.error.length > 0) {
       showError = <Text>{this.state.error}</Text>
+    }
+
+    let urlInvalid = '';
+    if (this.state.urlInvalid) {
+      urlInvalid = 'Please Enter a Valid URL';
+    }
+
+
+    let loginError = '';
+    if (this.state.loginError) {
+      loginError = this.state.loginErrorMessage;
     }
 
     return (
@@ -191,11 +226,26 @@ class LoginScreen extends React.Component {
                        onChangeText={(password) => this.setState({password})}/>
           </View>
 
+          <View>
+            <Text>{urlInvalid}</Text>
+          </View>
+
+
           <View style={styles.inputContainer}>
+
             <TextInput style={styles.inputs}
                        placeholder="Url"
                        underlineColorAndroid='transparent'
-                       onChangeText={(url) => this.setState({url})}/>
+                       onChangeText={
+                         (url) => {
+                           this.setState({url});
+                           this.setState({'urlInvalid': false})
+                         }
+                       }/>
+          </View>
+
+          <View>
+            <Text>{loginError}</Text>
           </View>
 
           <TouchableHighlight style={[styles.buttonContainer, styles.loginButton]}
